TST: simplify tests
