+++ examples/python/SRWLIB_Example16.py
@@ -9,7 +9,7 @@ of intensity distribution after diffraction on a circular aperture with an analy
 The example requires SciPy library to perform comparison.
 """
 from __future__ import print_function  # Python 2.7 compatibility
-import srwpy.uti_plot
+import srwpy.uti_plot as uti_plot
 from srwpy.srwlib import *
 from srwpy.uti_math import fwhm
